+++ src/layouts/MainLayout.vue
@@ -5,7 +5,3 @@
     </q-page-container>
   </q-layout>
 </template>
-
-<script setup>
-
-</script>